requirements.txt
aniso8601==8.0.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
flask-restplus==0.13.0
Flask-Testing==0.8.0
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
pyrsistent==0.16.0
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
Werkzeug==0.16.1
zipp==3.1.0
gunicorn==21.2.0
redis==5.0.0
psycopg2-binary
asyncpg==0.28.0